+++ .github/workflows/maven.yml
@@ -14,11 +14,11 @@ jobs:
       with:
         java-version: 1.8
     - name: Build with Maven
-      run: mvn -B package --file JoinMusic/pom.xml
+      run: mvn -B package --file pom.xml
     - name: Upload artifact
       uses: actions/upload-artifact@v1.0.0
       with:
         # Artifact name
         name: JoinMusic
         # Directory containing files to upload
-        path: JoinMusic/target/JoinMusic.jar
+        path: target/JoinMusic.jar
