--- JoinMusic/.gitignore
@@ -1,5 +1,0 @@
-/bin/
-/.settings/
-.classpath
-/.project
-/target/
