--- JoinMusic/.idea/libraries/Maven__com_github_koca2000_NoteBlockAPI_1_4_4.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: com.github.koca2000:NoteBlockAPI:1.4.4">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/github/koca2000/NoteBlockAPI/1.4.4/NoteBlockAPI-1.4.4.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/com/github/koca2000/NoteBlockAPI/1.4.4/NoteBlockAPI-1.4.4-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/github/koca2000/NoteBlockAPI/1.4.4/NoteBlockAPI-1.4.4-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
