--- JoinMusic/.idea/libraries/Maven__com_googlecode_json_simple_json_simple_1_1_1.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: com.googlecode.json-simple:json-simple:1.1.1">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/googlecode/json-simple/json-simple/1.1.1/json-simple-1.1.1.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/com/googlecode/json-simple/json-simple/1.1.1/json-simple-1.1.1-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/googlecode/json-simple/json-simple/1.1.1/json-simple-1.1.1-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
