--- JoinMusic/.idea/libraries/Maven__commons_lang_commons_lang_2_6.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: commons-lang:commons-lang:2.6">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/commons-lang/commons-lang/2.6/commons-lang-2.6.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/commons-lang/commons-lang/2.6/commons-lang-2.6-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/commons-lang/commons-lang/2.6/commons-lang-2.6-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
