--- JoinMusic/.idea/libraries/Maven__junit_junit_4_10.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: junit:junit:4.10">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/junit/junit/4.10/junit-4.10.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/junit/junit/4.10/junit-4.10-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/junit/junit/4.10/junit-4.10-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
