--- JoinMusic/.idea/libraries/Maven__org_bstats_bstats_bukkit_1_7.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.bstats:bstats-bukkit:1.7">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/bstats/bstats-bukkit/1.7/bstats-bukkit-1.7.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/bstats/bstats-bukkit/1.7/bstats-bukkit-1.7-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/bstats/bstats-bukkit/1.7/bstats-bukkit-1.7-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
