--- JoinMusic/.idea/libraries/Maven__org_yaml_snakeyaml_1_19.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.yaml:snakeyaml:1.19">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/yaml/snakeyaml/1.19/snakeyaml-1.19.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/yaml/snakeyaml/1.19/snakeyaml-1.19-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/yaml/snakeyaml/1.19/snakeyaml-1.19-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
