--- JoinMusic/src/plugin.yml
@@ -1,22 +1,0 @@
-name: JoinMusic
-authors:
-- T0biii
-- Fruxz
-depend:
-- NoteBlockAPI
-version: '4.0'
-website: http://dev.bukkit.org/bukkit-plugins/joinmusik
-main: de.t0biii.music.main.Main
-database: false
-commands:
-   JoinMusic:
-      description: /JoinMusic
-      usage: /<command> [reload]
-      aliases:
-      - jm
-      - joinmusic
-      - joinmusik
-      permissions: 
-        JoinMusik.command.reload
-        JoinMusik.command.stop
-        JoinMusik.use
